backend/requirements.txt
flask
flask-cors
flask-pymongo
pymongo
python-dotenv
